Improve error handling in new API
Errors from `roux` are forwarded through the stream, which now uses
`Result<SubmissionsData, RouxError>` as the item type. After some
experients I decided to just ignore errors from the sink: basically the
only thing that can happen with an unbounded channel is that we run out
of memory, an error for which we don't need to provide a complicated
infrastructure.

Retries still haven't been implemented for the new API.